+ Quản lý đơn hàng + Đặt hàng

+++ src/main/java/phuongnguyen/btl/entity/SelectedPhanLoaiDienThoai.java
@@ -4,6 +4,7 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -16,5 +17,5 @@ public class SelectedPhanLoaiDienThoai {
   MauSac[] selectedMauSacList;
   DungLuong[] selectedDungLuongList;
   Integer[] selectedSoLuong;
-  Double[] selectedGiaTien;
+  BigDecimal[] selectedGiaTien;
 }
